Refactor and Split Up Some Code
This refactors some code to use CSS modules, functional React components
and splits up the Settings component into roughly one component per
setting.
