+++ src/css/Layout.module.scss
@@ -1,0 +1,40 @@
+@use "Font";
+
+.resizableLayout {
+    position: absolute;
+    top: 0;
+    width: inherit;
+    height: inherit;
+
+    :global(.react-resizable) {
+        width: 0 !important;
+        height: 0 !important;
+    }
+}
+
+.handleEast {
+    cursor: e-resize;
+    right: -10px;
+    bottom: 10px;
+    top: 0;
+    position: absolute;
+    width: 20px;
+}
+
+.handleSouth {
+    cursor: s-resize;
+    bottom: -10px;
+    left: 0;
+    right: 10px;
+    position: absolute;
+    height: 20px;
+}
+
+.handleSouthEast {
+    cursor: se-resize;
+    bottom: -10px;
+    right: -10px;
+    position: absolute;
+    width: 20px;
+    height: 20px;
+}
