--- src/css/Layout.scss
@@ -1,42 +1,0 @@
-@use "Font";
-
-.layout {
-    .resizable-layout {
-        position: absolute;
-        top: 0;
-        width: inherit;
-        height: inherit;
-
-        .react-resizable {
-            width: 0 !important;
-            height: 0 !important;
-        }
-
-        .resizable-handle-east {
-            cursor: e-resize;
-            right: -10px;
-            bottom: 10px;
-            top: 0;
-            position: absolute;
-            width: 20px;
-        }
-
-        .resizable-handle-south {
-            cursor: s-resize;
-            bottom: -10px;
-            left: 0;
-            right: 10px;
-            position: absolute;
-            height: 20px;
-        }
-
-        .resizable-handle-south-east {
-            cursor: se-resize;
-            bottom: -10px;
-            right: -10px;
-            position: absolute;
-            width: 20px;
-            height: 20px;
-        }
-    }
-}
