+++ src/css/mobile.module.scss
@@ -1,0 +1,5 @@
+@mixin mobile {
+    :global(.is-mobile) & {
+        @content;
+    }
+}
