+++ src/ui/components/Dialog.tsx
@@ -1,11 +1,6 @@
-import * as React from "react";
+import React, { useState, useEffect } from "react";
 
-import "../../css/Dialog.scss";
-
-export interface Props {
-    onShow: () => void;
-    onClose: () => void;
-}
+import * as classes from "../../css/Dialog.module.scss";
 
 export interface Options {
     title: string | React.JSX.Element;
@@ -21,7 +16,7 @@ export interface State {
 }
 
 let dialogElement: HTMLDialogElement | null = null;
-let setState: ((options: Options) => void) | undefined;
+let setStateFn: ((options: Options) => void) | undefined;
 let resolveFn: ((_: [number, string]) => void) | undefined;
 let onCloseFn: (() => void) | undefined;
 let alreadyClosed = false;
@@ -39,99 +34,91 @@ export function showDialog(options: Options): Promise<[number, string]> {
             }
         };
     }
-    setState?.(options);
+    setStateFn?.(options);
     return new Promise((resolve) => (resolveFn = resolve));
 }
 
-export default class DialogContainer extends React.Component<Props, State> {
-    constructor(props: Props) {
-        super(props);
+export function DialogContainer({
+    onShow,
+    onClose,
+}: {
+    onShow: () => void;
+    onClose: () => void;
+}) {
+    const [options, setOptions] = useState<Options>({
+        title: "",
+        description: "",
+        buttons: [],
+    });
+    const [input, setInput] = useState("");
 
-        onCloseFn = props.onClose;
+    useEffect(() => {
+        onCloseFn = onClose;
+    }, [onClose]);
 
-        this.state = {
-            options: {
-                title: "",
-                description: "",
-                buttons: [],
-            },
-            input: "",
+    useEffect(() => {
+        setStateFn = (options) => {
+            onShow();
+            setOptions(options);
+            setInput(options.defaultText ?? "");
         };
-    }
+    }, [onShow]);
 
-    public componentDidMount(): void {
-        setState = (options) => {
-            this.props.onShow();
-            this.setState({
-                options,
-                input: options.defaultText ?? "",
-            });
-        };
-    }
-
-    public render() {
-        return (
-            <dialog
-                ref={(element) => {
-                    dialogElement = element;
-                }}
-                onKeyDown={(e) => {
-                    if (e?.key === "ArrowLeft") {
-                        e.preventDefault();
-                        (
-                            document.activeElement
-                                ?.previousElementSibling as any
-                        )?.focus();
-                    } else if (e?.key === "ArrowRight") {
-                        e.preventDefault();
-                        (
-                            document.activeElement?.nextElementSibling as any
-                        )?.focus();
-                    }
-                }}
-            >
-                <div className="dialog">
-                    <h1>{this.state.options.title}</h1>
-                    <p>{this.state.options.description}</p>
-                    {this.state.options.textInput && (
-                        <input
-                            type="text"
-                            value={this.state.input}
-                            autoFocus={true}
-                            onChange={(e) =>
-                                this.setState({ input: e.target.value })
-                            }
-                            onKeyDown={(e) => {
-                                if (e?.key === "Enter") {
-                                    e.preventDefault();
-                                    this.close(0);
-                                }
-                            }}
-                        />
-                    )}
-                    <div className="buttons">
-                        {this.state.options.buttons.map((button, i) => {
-                            return (
-                                <button
-                                    autoFocus={
-                                        i === 0 && !this.state.options.textInput
-                                    }
-                                    onClick={() => this.close(i)}
-                                >
-                                    {button}
-                                </button>
-                            );
-                        })}
-                    </div>
-                </div>
-            </dialog>
-        );
-    }
-
-    private close(i: number) {
+    const handleClose = (i: number) => {
         alreadyClosed = true;
         dialogElement?.close();
-        resolveFn?.([i, this.state.input]);
-        this.props.onClose();
-    }
+        resolveFn?.([i, input]);
+        onClose();
+    };
+
+    return (
+        <dialog
+            ref={(element) => {
+                dialogElement = element;
+            }}
+            onKeyDown={(e) => {
+                if (e?.key === "ArrowLeft") {
+                    e.preventDefault();
+                    (
+                        document.activeElement?.previousElementSibling as any
+                    )?.focus();
+                } else if (e?.key === "ArrowRight") {
+                    e.preventDefault();
+                    (
+                        document.activeElement?.nextElementSibling as any
+                    )?.focus();
+                }
+            }}
+        >
+            <div className={classes.dialog}>
+                <h1>{options.title}</h1>
+                <p>{options.description}</p>
+                {options.textInput && (
+                    <input
+                        type="text"
+                        value={input}
+                        autoFocus={true}
+                        onChange={(e) => setInput(e.target.value)}
+                        onKeyDown={(e) => {
+                            if (e?.key === "Enter") {
+                                e.preventDefault();
+                                handleClose(0);
+                            }
+                        }}
+                    />
+                )}
+                <div className={classes.buttons}>
+                    {options.buttons.map((button, i) => (
+                        <button
+                            key={i}
+                            autoFocus={i === 0 && !options.textInput}
+                            onClick={() => handleClose(i)}
+                        >
+                            {button}
+                        </button>
+                    ))}
+                </div>
+            </div>
+        </dialog>
+    );
 }
