+++ src/ui/components/DragUpload.tsx
@@ -1,37 +1,38 @@
-import * as React from "react";
+import React, { useEffect, useRef } from "react";
 import { toast } from "react-toastify";
 
-import "../../css/DragUpload.scss";
+import * as classes from "../../css/DragUpload.module.scss";
 
-export interface Props {
+export function DragUpload({
+    children,
+    importLayout,
+    importSplits,
+}: {
     children: React.ReactNode;
     importLayout?: (file: File) => Promise<void>;
-    importSplits(file: File): Promise<void>;
-}
+    importSplits: (file: File) => Promise<void>;
+}) {
+    const dropZoneRef = useRef<HTMLDivElement>(null);
+    const dropZoneOverlayRef = useRef<HTMLDivElement>(null);
 
-export default class DragUpload extends React.Component<Props> {
-    public componentDidMount() {
-        const dropZone = document.getElementById("upload-drop-zone");
-        const dropZoneOverlay = document.getElementById(
-            "upload-drop-zone-overlay",
-        );
-        const importLayout = this.props.importLayout;
-        const importSplits = this.props.importSplits;
+    useEffect(() => {
+        const dropZone = dropZoneRef.current;
+        const dropZoneOverlay = dropZoneOverlayRef.current;
 
-        if (dropZone === null) {
+        if (!dropZone) {
             return;
         }
 
-        dropZone.addEventListener("dragenter", (event) => {
+        const handleDragEnter = (event: DragEvent) => {
             event.preventDefault();
             event.stopPropagation();
 
             if (dropZoneOverlay) {
                 dropZoneOverlay.style.visibility = "visible";
             }
-        });
+        };
 
-        dropZone.addEventListener("dragleave", (event) => {
+        const handleDragLeave = (event: DragEvent) => {
             if (
                 dropZoneOverlay &&
                 (event.pageX < 10 ||
@@ -41,14 +42,14 @@ export default class DragUpload extends React.Component<Props> {
             ) {
                 dropZoneOverlay.style.visibility = "hidden";
             }
-        });
+        };
 
-        dropZone.addEventListener("dragover", (event) => {
+        const handleDragOver = (event: DragEvent) => {
             event.preventDefault();
             event.stopPropagation();
-        });
+        };
 
-        dropZone.addEventListener("drop", (event) => {
+        const handleDrop = (event: DragEvent) => {
             event.preventDefault();
             event.stopPropagation();
 
@@ -73,17 +74,27 @@ export default class DragUpload extends React.Component<Props> {
                 }
             }
             return null;
-        });
-    }
+        };
 
-    public render() {
-        return (
-            <div id="upload-drop-zone">
-                <div id="upload-drop-zone-overlay">
-                    <div className="overlay-text">Waiting for drop...</div>
-                </div>
-                {this.props.children}
+        dropZone.addEventListener("dragenter", handleDragEnter);
+        dropZone.addEventListener("dragleave", handleDragLeave);
+        dropZone.addEventListener("dragover", handleDragOver);
+        dropZone.addEventListener("drop", handleDrop);
+
+        return () => {
+            dropZone.removeEventListener("dragenter", handleDragEnter);
+            dropZone.removeEventListener("dragleave", handleDragLeave);
+            dropZone.removeEventListener("dragover", handleDragOver);
+            dropZone.removeEventListener("drop", handleDrop);
+        };
+    }, [importLayout, importSplits]);
+
+    return (
+        <div ref={dropZoneRef} className={classes.uploadDropZone}>
+            <div ref={dropZoneOverlayRef} className={classes.overlay}>
+                <div className={classes.overlayText}>Waiting for drop...</div>
             </div>
-        );
-    }
+            {children}
+        </div>
+    );
 }
