+++ src/ui/components/Layout.tsx
@@ -6,9 +6,18 @@ import AutoRefresh from "../../util/AutoRefresh";
 import { UrlCache } from "../../util/UrlCache";
 import { GeneralSettings } from "../views/MainSettings";
 
-import "../../css/Layout.scss";
+import * as classes from "../../css/Layout.module.scss";
 
-export interface Props {
+export default function Layout({
+    getState,
+    layoutUrlCache,
+    allowResize,
+    width,
+    height,
+    generalSettings,
+    renderer,
+    onResize,
+}: {
     getState: () => LayoutStateRef;
     layoutUrlCache: UrlCache;
     allowResize: boolean;
@@ -16,100 +25,78 @@ export interface Props {
     height: number;
     generalSettings: GeneralSettings;
     renderer: WebRenderer;
-    onResize(width: number, height: number): void;
-}
-
-export default class Layout extends React.Component<Props, unknown> {
-    public refreshLayout() {
-        const layoutState = this.props.getState();
-        const newDims = this.props.renderer.render(
+    onResize: (width: number, height: number) => void;
+}) {
+    const update = () => {
+        const layoutState = getState();
+        const newDims = renderer.render(
             layoutState.ptr,
-            this.props.layoutUrlCache.imageCache.ptr,
+            layoutUrlCache.imageCache.ptr,
         );
-        if (newDims !== undefined) {
-            this.props.onResize(newDims[0], newDims[1]);
+        if (newDims != null) {
+            onResize(newDims[0], newDims[1]);
         }
-    }
+    };
 
-    public render() {
-        return (
-            <AutoRefresh
-                frameRate={this.props.generalSettings.frameRate}
-                update={() => this.refreshLayout()}
-            >
+    return (
+        <AutoRefresh frameRate={generalSettings.frameRate} update={update}>
+            <div style={{ width, height }}>
                 <div
-                    className="layout"
-                    style={{
-                        width: this.props.width,
-                        height: this.props.height,
+                    style={{ width, height }}
+                    ref={(element) => {
+                        element?.appendChild(renderer.element());
                     }}
-                >
-                    <div
-                        style={{ width: "inherit", height: "inherit" }}
-                        ref={(element) => {
-                            element?.appendChild(this.props.renderer.element());
-                        }}
-                    />
-                    {this.props.allowResize && (
-                        <div className="resizable-layout">
-                            <ResizableBox
-                                axis="x"
-                                width={this.props.width}
-                                height={this.props.height}
-                                minConstraints={[100, 40]}
-                                handle={
-                                    <div
-                                        onClick={(e) => e.stopPropagation()}
-                                        className="resizable-handle-east"
-                                    />
-                                }
-                                onResize={(_event, data) =>
-                                    this.props.onResize(
-                                        data.size.width,
-                                        data.size.height,
-                                    )
-                                }
-                            />
-                            <ResizableBox
-                                axis="y"
-                                width={this.props.width}
-                                height={this.props.height}
-                                minConstraints={[100, 40]}
-                                handle={
-                                    <div
-                                        onClick={(e) => e.stopPropagation()}
-                                        className="resizable-handle-south"
-                                    />
-                                }
-                                onResize={(_event, data) =>
-                                    this.props.onResize(
-                                        data.size.width,
-                                        data.size.height,
-                                    )
-                                }
-                            />
-                            <ResizableBox
-                                axis="both"
-                                width={this.props.width}
-                                height={this.props.height}
-                                minConstraints={[100, 40]}
-                                handle={
-                                    <div
-                                        onClick={(e) => e.stopPropagation()}
-                                        className="resizable-handle-south-east"
-                                    />
-                                }
-                                onResize={(_event, data) =>
-                                    this.props.onResize(
-                                        data.size.width,
-                                        data.size.height,
-                                    )
-                                }
-                            />
-                        </div>
-                    )}
-                </div>
-            </AutoRefresh>
-        );
-    }
+                />
+                {allowResize && (
+                    <div className={classes.resizableLayout}>
+                        <ResizableBox
+                            axis="x"
+                            width={width}
+                            height={height}
+                            minConstraints={[100, 40]}
+                            handle={
+                                <div
+                                    onClick={(e) => e.stopPropagation()}
+                                    className={classes.handleEast}
+                                />
+                            }
+                            onResize={(_event, data) =>
+                                onResize(data.size.width, data.size.height)
+                            }
+                        />
+                        <ResizableBox
+                            axis="y"
+                            width={width}
+                            height={height}
+                            minConstraints={[100, 40]}
+                            handle={
+                                <div
+                                    onClick={(e) => e.stopPropagation()}
+                                    className={classes.handleSouth}
+                                />
+                            }
+                            onResize={(_event, data) =>
+                                onResize(data.size.width, data.size.height)
+                            }
+                        />
+                        <ResizableBox
+                            axis="both"
+                            width={width}
+                            height={height}
+                            minConstraints={[100, 40]}
+                            handle={
+                                <div
+                                    onClick={(e) => e.stopPropagation()}
+                                    className={classes.handleSouthEast}
+                                />
+                            }
+                            onResize={(_event, data) =>
+                                onResize(data.size.width, data.size.height)
+                            }
+                        />
+                    </div>
+                )}
+            </div>
+        </AutoRefresh>
+    );
 }
