+++ src/ui/views/About.tsx
@@ -5,11 +5,7 @@ import { Markdown } from "../../util/Markdown";
 import { ArrowLeft } from "lucide-react";
 
 import * as variables from "../../css/variables.icss.scss";
-import "../../css/About.scss";
-
-export interface Props {
-    callbacks: Callbacks;
-}
+import * as classes from "../../css/About.module.scss";
 
 interface Callbacks {
     renderViewWithSidebar(
@@ -21,96 +17,90 @@ interface Callbacks {
 
 const contributorAvatarSize = parseFloat(variables.contributorAvatarSize);
 
-export class About extends React.Component<Props> {
-    public render() {
-        const renderedView = this.renderView();
-        const sidebarContent = this.renderSidebarContent();
-        return this.props.callbacks.renderViewWithSidebar(
-            renderedView,
-            sidebarContent,
-        );
-    }
+export function About({ callbacks }: { callbacks: Callbacks }) {
+    const renderedView = renderView();
+    const sidebarContent = renderSidebarContent(callbacks);
+    return callbacks.renderViewWithSidebar(renderedView, sidebarContent);
+}
 
-    private renderView() {
-        const idealAvatarResolution = Math.round(
-            devicePixelRatio * contributorAvatarSize,
-        );
+function renderView() {
+    const idealAvatarResolution = Math.round(
+        devicePixelRatio * contributorAvatarSize,
+    );
 
-        return (
-            <div className="about">
-                <div className="about-inner-container">
-                    <div className="livesplit-title">
-                        <span className="livesplit-icon">
-                            <img src={LiveSplitIcon} alt="LiveSplit Logo" />
-                        </span>
-                        <div className="title-text">LiveSplit One</div>
-                    </div>
-                    <p className="build-version">
-                        <a
-                            href={`https://github.com/LiveSplit/LiveSplitOne/commit/${COMMIT_HASH}`}
-                            target="_blank"
-                        >
-                            Version: {BUILD_DATE}
-                        </a>
-                    </p>
-                    <p>
-                        LiveSplit One is a multiplatform version of LiveSplit,
-                        the sleek, highly-customizable timer for speedrunners.
-                    </p>
-                    <p>
-                        <a
-                            href="https://github.com/LiveSplit/LiveSplitOne"
-                            target="_blank"
-                        >
-                            View Source Code on GitHub
-                        </a>
-                    </p>
-                    <h2>Recent Changes</h2>
-                    <div className="changelog">
-                        {CHANGELOG.map((change) => (
-                            <>
-                                <a
-                                    href={`https://github.com/LiveSplit/LiveSplitOne/commit/${change.id}`}
-                                    target="_blank"
-                                >
-                                    {change.date}
-                                </a>
-                                <Markdown
-                                    markdown={change.message}
-                                    unsafe={true}
-                                />
-                            </>
-                        ))}
-                    </div>
-                    <h2>Contributors</h2>
-                    <div className="contributors">
-                        {CONTRIBUTORS_LIST.map((contributor) => (
+    return (
+        <div className={classes.about}>
+            <div className={classes.aboutInnerContainer}>
+                <div className={classes.livesplitTitle}>
+                    <img
+                        className={classes.livesplitIcon}
+                        src={LiveSplitIcon}
+                        alt="LiveSplit Logo"
+                    />
+                    <div className={classes.titleText}>LiveSplit One</div>
+                </div>
+                <p className={classes.buildVersion}>
+                    <a
+                        href={`https://github.com/LiveSplit/LiveSplitOne/commit/${COMMIT_HASH}`}
+                        target="_blank"
+                    >
+                        Version: {BUILD_DATE}
+                    </a>
+                </p>
+                <p>
+                    LiveSplit One is a multiplatform version of LiveSplit, the
+                    sleek, highly-customizable timer for speedrunners.
+                </p>
+                <p>
+                    <a
+                        href="https://github.com/LiveSplit/LiveSplitOne"
+                        target="_blank"
+                    >
+                        View Source Code on GitHub
+                    </a>
+                </p>
+                <h2>Recent Changes</h2>
+                <div className={classes.changelog}>
+                    {CHANGELOG.map((change) => (
+                        <>
                             <a
-                                href={`https://github.com/${contributor.name}`}
+                                href={`https://github.com/LiveSplit/LiveSplitOne/commit/${change.id}`}
                                 target="_blank"
                             >
-                                <img
-                                    src={`https://avatars.githubusercontent.com/u/${contributor.id}?s=${idealAvatarResolution}&v=4`}
-                                    onError={(e) => (e.target as any).remove()}
-                                />
-                                {contributor.name}
+                                {change.date}
                             </a>
-                        ))}
-                    </div>
+                            <Markdown markdown={change.message} unsafe={true} />
+                        </>
+                    ))}
+                </div>
+                <h2>Contributors</h2>
+                <div className={classes.contributors}>
+                    {CONTRIBUTORS_LIST.map((contributor) => (
+                        <a
+                            href={`https://github.com/${contributor.name}`}
+                            target="_blank"
+                        >
+                            <img
+                                src={`https://avatars.githubusercontent.com/u/${contributor.id}?s=${idealAvatarResolution}&v=4`}
+                                onError={(e) => (e.target as any).remove()}
+                            />
+                            {contributor.name}
+                        </a>
+                    ))}
                 </div>
             </div>
-        );
-    }
+        </div>
+    );
+}
 
-    private renderSidebarContent() {
-        return (
-            <div className="sidebar-buttons">
-                <h1>About</h1>
-                <hr />
-                <button onClick={(_) => this.props.callbacks.openTimerView()}>
-                    <ArrowLeft strokeWidth={2.5} /> Back
-                </button>
-            </div>
-        );
-    }
+function renderSidebarContent(callbacks: Callbacks) {
+    return (
+        <div className="sidebar-buttons">
+            <h1>About</h1>
+            <hr />
+            <button onClick={(_) => callbacks.openTimerView()}>
+                <ArrowLeft strokeWidth={2.5} /> Back
+            </button>
+        </div>
+    );
 }
