+++ src/util/OptionUtil.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import { toast } from "react-toastify";
 
 export type Option<T> = T | null | undefined;
